Fix failing pacsort tests
In commit 42181f85c44de29d40642aeed897e030f30e0559 the pacsort testsuite
was migrated from the pacman codebase, and the PMTEST_UTIL_DIR variable
was changed to PMTEST_BIN_DIR. However, it was only changed in two of
the three places it was needed, and the testsuite could not find the
pacsort binary in the build tree.

As a result, the tests would fail on a system that did not have a
systemwide pacsort install, and would test the systemwide install if
there was one available.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>

Message-Id: <20171113171454.24555-2-eschwartz@archlinux.org>
Signed-off-by: Johannes Löthberg <johannes@kyriasis.com>