pacscripts: use filename directly instead of parsing pacman.conf
pacman-conf may return multiple cachedirs, in which case we treat it as
one cachedir containing newline characters, which is certainly wrong in
every case. However, pacman -Sddp already returns a filepath when the
package is cached. This also properly handles packages already available
in a secondary cache directory without looping through each one and
checking for ourselves, where it is located.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>
Signed-off-by: Daniel M. Capella <polycitizen@gmail.com>