paclist: list packages which version differs from the repo version
Paclist does not list installed packages if the version is different
from the version in the repository. Change the awk command to also
match such packages from the 'pacman -Sl' output and print the
installed version. Also paclist fails for languages which translate
'installed' to more than one word. So do not use the field numbers
in a record for matching.

Fixes FS#60412

Signed-off-by: Michael Straube <michael.straube@posteo.de>