autotools: remove useless file from the source tree
tap-driver.sh is a generated file, it is installed by autoreconf because
configure.ac says to install it. There's no reason to keep a copy
checked into git, and by tracking it, users end up with dirty trees just
because they attempted to build the project. This prevents convenient
use of various git features.

Its presence in git is most likely an accident.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>
Message-Id: <20190417024631.19259-1-eschwartz@archlinux.org>
Signed-off-by: Johannes Löthberg <johannes@kyriasis.com>