updpkgsums: rely on makepkg's exported schema declaration for algo types
Rather than heuristically match (md|sha)[[:digit:]]+sums or even .*sums,
read makepkg's internal description of known hash algorithms, and use it
to generate a search pattern. This ensures that we will always be in
sync with the makepkg installed on the user's system.

Fixes support for the newly added b2sums algorithm.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>