build: re-add dependency link between paccache and size_to_human.sh
Fixes regression in commit 431e564ed99dcd774e35b0232435e1d6c7f0c30b,
which ported most of lib/ to libmakepkg extensions, but size_to_human is
not part of libmakepkg. An initial version of the patch removed all of
lib/, but when I realized this broke paccache, I re-added size_to_human
to the tree. However, I neglected to add the Makefile dependency back at
the same time, before sending in a 2 AM patchset, and by the time I
noticed this the next morning, it was already merged...

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>