checkupdates: use $UID instead of $USER in temporary directory path
If $USER is not set when running checkupdates, the temporary directory's
name will be "checkup-db-". This will cause issues if multiple users run
checkupdates. A common situation where $USER is not set is when a command
is executed with systemd-run, e.g. from a systemd unit.
See https://github.com/systemd/systemd/pull/8227 for details.

Replace the usage of $USER with $UID since it is guaranteed to exist and so
that the temporary directory is isolated per-user.

Signed-off-by: nl6720 <nl6720@gmail.com>
Message-Id: <20191120165707.34383-1-nl6720@gmail.com>
Signed-off-by: Johannes Löthberg <johannes@kyriasis.com>