paccache: Support cleaning many thousands of candidates
In situations when there are many thousands of candidates for deletion,
paccache returns an "Argument list too long" error at the freed space
calculation stage, because the expansion of the ${candidates[@]} array
becomes too long for Bash.

This commit fixes that problem, by getting the @SIZECMD@ arguments via
`printf '%s\0' "${candidates[@]}" | xargs -0`, similarly to what is done
for the (re)moving commands below.

Signed-off-by: Leonid Bloch <lb.workbox@gmail.com>
Signed-off-by: Daniel M. Capella <polyzen@archlinux.org>