pactree: Hide dependency satisfiers in more cases
Pactree takes care not to print both a dependency and the package to
which it resolves if they are equal. Extend (fix?) this to cover
versioned dependencies and optional dependencies containing
descriptions, so that e.g. `gnupg provides gnupg>=2` and `clang provides
clang: for qdoc` will be made more compact.

Implementation suggested by Eli Schwartz.

Signed-off-by: Sebastian Jakubiak <jakseb.dev@gmail.com>
Signed-off-by: Daniel M. Capella <polyzen@archlinux.org>