updpkgsums: don't try to add nonexistent checksums
In the event there are no sources, there won't be sums either. awk won't
find any to remove and replace inline, so it tries to print "" at the
end of the file, and merely adds superfluous newlines.

Abort early instead, since updpkgsums cannot be expected to process such
a file.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>
Signed-off-by: Daniel M. Capella <polyzen@archlinux.org>