pacdiff: add option to view diffs in 3-way fashion
add -3/--threeway option to view diffs in 3-way fashion.
this is similar to the 3-way merge, but instead of merging,
it opens all three files in $diffprog.

if the previous package version of the candidate is not available
in pacman's cache or if the file fails to extract, it falls back to
regular 2-way diff.